facecheck_requirements.txt
# FaceCheck Flask App + Telegram Bot Requirements
flask>=2.0.0
flask-httpauth>=4.8.0
werkzeug>=2.3.0
requests>=2.25.0
python-telegram-bot>=20.0
python-dotenv>=0.19.0
gunicorn>=20.1.0




